requirements.txt
google-cloud-pubsub~=2.13.6
pydantic>=1.8.0,<2.0.0
cryptography==3.3.2
munch